feat(network): save user selected network
