+++ client/src/common/hooks/useTheme.ts
@@ -1,26 +1,5 @@
 import { useEffect, useState } from 'react'
-
-function useSafeLocalStorage(key, initialValue) {
-  const [valueProxy, setValueProxy] = useState(() => {
-    try {
-      const value = window.localStorage.getItem(key)
-      return value ? JSON.parse(value) : initialValue
-    } catch {
-      return initialValue
-    }
-  })
-
-  const setValue = value => {
-    try {
-      window.localStorage.setItem(key, value)
-      setValueProxy(value)
-    } catch {
-      setValueProxy(value)
-    }
-  }
-
-  return [valueProxy, setValue]
-}
+import { useSafeLocalStorage } from './useSafeLocalStorage'
 
 function usePrefersDarkMode() {
   const [value, setValue] = useState(true)
